+++ data/libs/systems/features.lua
@@ -2,6 +2,7 @@ Features = {
 	AutoLoot = "autoloot",
 	ChainSystem = "chainSystem",
 	EmoteSpells = "emoteSpells",
+	SpellNameInsteadOfWords = "spellNameInsteadOfWords",
 }
 
 local function validateFeature(feature)