+++ src/creatures/players/player.hpp
@@ -1212,6 +1212,7 @@ public:
 	bool checkAutoLoot(bool isBoss) const;
 	bool checkChainSystem() const;
 	bool checkEmoteSpells() const;
+	bool checkSpellNameInsteadOfWords() const;
 
 	QuickLootFilter_t getQuickLootFilter() const;